[Update] 날씨 관련된 값들 Long이 아니라 Double 타입으로 변경

+++ src/main/java/Nanali/domain/cody/cloth/Garment.java
@@ -27,12 +27,12 @@ public class Garment extends BaseEntity { // Garment는 코디 안의 아이템 
     @Enumerated(EnumType.STRING)
     private Sex sex;
 
-    private Long temp_from;
-    private Long temp_to;
-    private Long uv_from;
-    private Long uv_to;
-    private Long rain_from;
-    private Long rain_to;
+    private Double temp_from;
+    private Double temp_to;
+    private Double uv_from;
+    private Double uv_to;
+    private Double rain_from;
+    private Double rain_to;
 
     @OneToMany(mappedBy = "garment")
     private List<LikeGarment> garmentList;
@@ -44,8 +44,8 @@ public class Garment extends BaseEntity { // Garment는 코디 안의 아이템 
 
     }
 
-    public Garment(String imgName, String imgUrl, Category category, Sex sex, Long temp_from, Long temp_to,
-                   Long uv_from, Long uv_to, Long rain_from, Long rain_to) {
+    public Garment(String imgName, String imgUrl, Category category, Sex sex, Double temp_from, Double temp_to,
+                   Double uv_from, Double uv_to, Double rain_from, Double rain_to) {
         this.imgName = imgName;
         this.imgUrl = imgUrl;
         this.category = category;
