+++ src/main/java/Nanali/dtos/weather/GarmentWeatherRequest.java
@@ -9,10 +9,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class GarmentWeatherRequest {
 
-    private Long temp_from;
-    private Long temp_to;
-    private Long uv_from;
-    private Long uv_to;
-    private Long rain_from;
-    private Long rain_to;
+    private Double temp_from;
+    private Double temp_to;
+    private Double uv_from;
+    private Double uv_to;
+    private Double rain_from;
+    private Double rain_to;
 }
